FIX: compiler warning

+++ src/paks/ejs/ejsLib.c
@@ -9920,7 +9920,7 @@ static int addCharToToken(EcToken *tp, int c)
     
     if (tp->length >= (tp->size - 1)) {
         boost = max(tp->size, EC_TOKEN_INCR);
-        tp->size += boost;
+        tp->size += (int) boost;
         if ((tp->text = mprRealloc(tp->text, tp->size * sizeof(wchar))) == 0) {
             return MPR_ERR_MEMORY;
         }
